optimize: 业务变量和服务变量相关接口分别放在不同文件
